requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
greenlet==1.1.1
kazoo==2.8.0
mysqlclient==2.0.3
pykafka==2.8.0
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
tabulate==0.8.9
